backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.4
python-dotenv==1.1.0
httpx==0.28.1
anthropic>=0.52.0
pandas>=2.2.0
